update to use safeParse to validate export options
