+++ app/(dashboard)/dashboard/interviews/_actions/export.ts
@@ -7,7 +7,19 @@ import { formatExportableSessions } from '~/lib/network-exporters/formatters/for
 import { api } from '~/trpc/server';
 import { getServerSession } from '~/utils/auth';
 import { ensureError } from '~/utils/ensureError';
-import { type ExportOptions } from '~/utils/validateExportOptions';
+import { z } from 'zod';
+
+export const ExportOptionsSchema = z.object({
+  exportGraphML: z.boolean(),
+  exportCSV: z.boolean(),
+  globalOptions: z.object({
+    exportFilename: z.string(),
+    unifyNetworks: z.boolean(),
+    useScreenLayoutCoordinates: z.boolean(),
+  }),
+});
+
+export type ExportOptions = z.infer<typeof ExportOptionsSchema>;
 
 type UploadData = {
   key: string;
